move map route
